requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-admin-soft-dashboard==1.0.12
django-jazzmin==2.6.0
idna==3.4
lxml==4.9.2
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
